plataforma/backend-fastapi/pyproject.toml
[project]
name = "ciber-platform-api"
version = "1.0.0"
description = "Multi-course cybersecurity LMS API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "PyJWT[crypto]>=2.8.0",
    "bcrypt>=4.0.0",
    "redis[hiredis]>=5.2.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "aiosmtplib>=3.0.2",
    "slowapi>=0.1.9",
    "structlog>=24.4.0",
    "python-magic>=0.4.27",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "factory-boy>=3.3.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*", "app/scripts/*"]

[tool.coverage.report]
fail_under = 65
show_missing = true
